+++ src/index.js
@@ -1,3 +1,1 @@
 export { default } from "./checkable";
-export { default as mapStore } from "./mapStore";
-export { default as stateStore } from "./stateStore";